// services/deploy/src/config/types.ts
export enum DbTables {
  BACKEND = 'backend',
}

export enum DeployErrorCode {
  //400
  //404,
  DEPLOY_NOT_FOUND = 40421001,
  DEPLOY_POD_NOT_FOUND = 40421002,
  //405

  //422
  DATA_NOT_PRESENT = 42221001,
  DATA_NOT_VALID = 42221002,

  //500
  GENERAL_SERVER_ERROR = 50021000,
  DEPLOY_DEPLOY_FAILED = 50021001,
  DEPLOY_RESIZE_FAILED = 50021002,
  // 501
}